fix: import models as dynamic
